Bugfix
QoS_3 was broken: assumed negative values. This was due to an EV charging at >13kW. If towards the end of its stay the updated predictions indicate that it needs access to 22kW charging, we do that. Generically fixed the QoS bug here by checking for the largest value in the power profile. If it's >13, we set the considered max to 22.