Bugfix
nx.min_cost_flow() does not always return a solution. If so, it may compute for a long time. We introduce a timeout to skip the LYNCS part of the scheduling and default back to the naive FOCS schedule, if computing takes longer than 10 seconds.

Hard coded for now.
FIXME: make timeout length dependent on input size.